+++ lib/rgeo/coord_sys.rb
@@ -24,14 +24,6 @@ module RGeo
   # systems, and other related concepts, as well as a parser for the WKT
   # format for specifying coordinate systems.
   module CoordSys
-    # The only valid key is :proj4
-    def self.supported?(key)
-      raise(Error::UnsupportedOperation, "Invalid key. The only valid key is :proj4.") unless key == :proj4
-      defined?(RGeo::CoordSys::Proj4) && RGeo::CoordSys::Proj4.supported?
-    end
-
-    def self.check!(key)
-      supported?(key) || raise(Error::UnsupportedOperation, "Coordinate system '#{key}' is not supported.")
-    end
+    CONFIG = Struct.new(:default_coord_sys_class).new(CS::CoordinateSystem)
   end
 end