+++ lib/rgeo/coord_sys/cs/entities.rb
@@ -939,6 +939,51 @@ module RGeo
         def get_units(dimension)
           nil
         end
+
+        def geographic?
+          false
+        end
+
+        def projected?
+          false
+        end
+
+        def wkt_typename
+          "CS"
+        end
+
+        # Not an OGC method, but useful for being able to
+        # transform directly from a CoordinateSystem object.
+        def transform_coords(target_cs, x, y, z = nil)
+          ct = CoordinateTransform.create(self, target_cs)
+          ct.transform_coords(x, y, z)
+        end
+
+        class << self
+          def create(defn, dimension = 2, *optional)
+            # Need this so we can maintain consistency with actual
+            # CoordinateSystem implementations
+
+            if defn.is_a?(Integer)
+              # not technically correct but we can use cartesian as a placeholder
+              # to form valid wkt
+              defn_string = "Cartesian"
+              new(defn_string, dimension, "EPSG", defn, *optional)
+            else
+              new(defn, dimension, *optional)
+            end
+          end
+
+          def create_from_wkt(str)
+            CS.create_from_wkt(str)
+          end
+        end
+
+        private
+
+        def wkt_content(_)
+          [@dimension]
+        end
       end
 
       # == OGC spec description
@@ -1108,6 +1153,10 @@ module RGeo
           [@axis0, @axis1, @axis2][index]
         end
 
+        def geographic?
+          true
+        end
+
         def wkt_typename
           "GEOCCS"
         end
@@ -1267,6 +1316,10 @@ module RGeo
           @horizontal_datum.wgs84_parameters
         end
 
+        def geographic?
+          true
+        end
+
         def wkt_typename
           "GEOGCS"
         end
@@ -1331,6 +1384,10 @@ module RGeo
 
         def get_axis(index)
           index == 1 ? @axis1 : @axis0
+        end
+
+        def projected?
+          true
         end
 
         def wkt_typename